Organise shared code relevant to applications
