+++ src/js/shared/apply/checkbox.js
