+++ src/js/shared/apply/render-table-form.js
